# docs/swagger/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the FeatherWings daemon.",
        "title": "FeatherWings API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/diagnostics": {
            "get": {
                "security": [
                    {
                        "NodeToken": []
                    }
                ],
                "description": "Returns plain-text diagnostics output by default. Use include_endpoints to append HTTP endpoint metadata and include_logs to attach recent logs. Provide format=url to upload the report and receive a shortened URL instead of the raw content.",
                "produces": [
                    "text/plain",
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Generate diagnostics bundle",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Include endpoint metadata",
                        "name": "include_endpoints",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include daemon logs",
                        "name": "include_logs",
                        "in": "query"
                    },
                    {
                        "maximum": 500,
                        "minimum": 1,
                        "type": "integer",
                        "description": "Number of log lines",
                        "name": "log_lines",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "text",
                            "url"
                        ],
                        "type": "string",
                        "description": "Response format",
                        "name": "format",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Override upload endpoint when format=url",
                        "name": "upload_api_url",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plain-text diagnostics report. When format=url the response type is application/json with payload {\\\"url\\\":\\\"...\\\"}.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/servers": {
            "get": {
                "security": [
                    {
                        "NodeToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Servers"
                ],
                "summary": "List servers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/server.APIResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "NodeToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Servers"
                ],
                "summary": "Create server",
                "parameters": [
                    {
                        "description": "Server configuration",
                        "name": "server",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/installer.ServerDetails"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/system": {
            "get": {
                "security": [
                    {
                        "NodeToken": []
                    }
                ],
                "description": "Returns hardware and software information about the node. Provide `v=2` to receive the full payload used by the panel.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Get system information",
                "parameters": [
                    {
                        "enum": [
                            "2"
                        ],
                        "type": "string",
                        "description": "Response version",
                        "name": "v",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Extended response when v=2",
                        "schema": {
                            "$ref": "#/definitions/system.Information"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/system/docker/disk": {
            "get": {
                "security": [
                    {
                        "NodeToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Get Docker disk usage",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/system.DockerDiskUsage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/system/docker/image/prune": {
            "delete": {
                "security": [
                    {
                        "NodeToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Prune Docker images",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/router.DockerPruneReport"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/system/ips": {
            "get": {
                "security": [
                    {
                        "NodeToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "List system IP addresses",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/system.IpAddresses"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/system/utilization": {
            "get": {
                "security": [
                    {
                        "NodeToken": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Get system utilization",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/system.Utilization"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/update": {
            "post": {
                "security": [
                    {
                        "NodeToken": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Update runtime configuration",
                "parameters": [
                    {
                        "description": "Updated configuration",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/config.Configuration"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/router.postUpdateConfigurationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "config.Configuration": {
            "type": "object"
        },
...
        "environment.Variables": {
            "type": "object",
            "additionalProperties": true
        },
...
        "installer.ServerDetails": {
            "type": "object",
            "properties": {
                "start_on_completion": {
                    "type": "boolean"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "router.DockerPruneReport": {
            "type": "object",
            "properties": {
                "imagesDeleted": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/image.DeleteResponse"
                    }
                },
                "spaceReclaimed": {
                    "type": "integer"
                }
            }
        },
        "router.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
        "router.postUpdateConfigurationResponse": {
            "type": "object",
            "properties": {
                "applied": {
                    "type": "boolean"
                }
            }
        },
        "server.APIResponse": {
            "type": "object",
            "properties": {
                "configuration": {
                    "$ref": "#/definitions/server.Configuration"
                },
                "is_suspended": {
                    "type": "boolean"
                },
                "state": {
                    "type": "string"
                },
                "utilization": {
                    "$ref": "#/definitions/server.ResourceUsage"
                }
            }
        },
        "server.Configuration": {
            "type": "object",
            "properties": {
                "allocations": {
                    "$ref": "#/definitions/environment.Allocations"
                },
                "build": {
                    "$ref": "#/definitions/environment.Limits"
                },
                "container": {
                    "type": "object",
                    "properties": {
                        "image": {
                            "description": "Defines the Docker image that will be used for this server",
                            "type": "string"
                        }
                    }
                },
                "crash_detection_enabled": {
                    "type": "boolean"
                },
                "egg": {
                    "$ref": "#/definitions/server.EggConfiguration"
                },
                "environment": {
                    "description": "An array of environment variables that should be passed along to the running\nserver process.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/environment.Variables"
                        }
                    ]
                },
                "invocation": {
                    "description": "The command that should be used when booting up the server instance.",
                    "type": "string"
                },
                "labels": {
                    "description": "Labels is a map of container labels that should be applied to the running server process.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/server.ConfigurationMeta"
                },
                "mounts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.Mount"
                    }
                },
                "skip_egg_scripts": {
                    "description": "By default this is false, however if selected within the Panel while installing or re-installing a\nserver, specific installation scripts will be skipped for the server process.",
                    "type": "boolean"
                },
                "suspended": {
                    "description": "Whether or not the server is in a suspended state. Suspended servers cannot\nbe started or modified except in certain scenarios by an admin user.",
                    "type": "boolean"
                },
                "uuid": {
                    "description": "The unique identifier for the server that should be used when referencing\nit against the Panel API (and internally). This will be used when naming\ndocker containers as well as in log output.",
                    "type": "string"
                }
            }
        },
...
        "server.ResourceUsage": {
            "type": "object",
            "properties": {
                "cpu_absolute": {
                    "description": "The absolute CPU usage is the amount of CPU used in relation to the entire system and\ndoes not take into account any limits on the server process itself.",
                    "type": "number"
                },
                "disk_bytes": {
                    "description": "The current disk space being used by the server. This value is not guaranteed to be accurate\nat all times. It is \"manually\" set whenever server.Proc() is called. This is kind of just a\nhacky solution for now to avoid passing events all over the place.",
                    "type": "integer"
                },
                "memory_bytes": {
                    "description": "The total amount of memory, in bytes, that this server instance is consuming. This is\ncalculated slightly differently than just using the raw Memory field that the stats\nreturn from the container, so please check the code setting this value for how that\nis calculated.",
                    "type": "integer"
                },
                "memory_limit_bytes": {
                    "description": "The total amount of memory this container or resource can use. Inside Docker this is\ngoing to be higher than you'd expect because we're automatically allocating overhead\nabilities for the container, so it's not going to be a perfect match.",
                    "type": "integer"
                },
                "network": {
                    "description": "Current network transmit in \u0026 out for a container.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/environment.NetworkStats"
                        }
                    ]
                },
                "state": {
                    "description": "The current server status.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/system.AtomicString"
                        }
                    ]
                },
                "uptime": {
                    "description": "The current uptime of the container, in milliseconds.",
                    "type": "integer"
                }
            }
        },
        "system.AtomicString": {
            "type": "object"
        },
...
        "system.DockerDiskUsage": {
            "type": "object",
            "properties": {
                "build_cache_size": {
                    "type": "integer"
                },
                "containers_size": {
                    "type": "integer"
                },
                "images_active": {
                    "type": "integer"
                },
                "images_size": {
                    "type": "integer"
                },
                "images_total": {
                    "type": "integer"
                }
            }
        },
        "system.DockerInformation": {
            "type": "object",
            "properties": {
                "cgroups": {
                    "$ref": "#/definitions/system.DockerCgroups"
                },
                "containers": {
                    "$ref": "#/definitions/system.DockerContainers"
                },
                "runc": {
                    "$ref": "#/definitions/system.DockerRunc"
                },
                "storage": {
                    "$ref": "#/definitions/system.DockerStorage"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "system.Information": {
            "type": "object",
            "properties": {
                "docker": {
                    "$ref": "#/definitions/system.DockerInformation"
                },
                "system": {
                    "$ref": "#/definitions/system.System"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "system.IpAddresses": {
            "type": "object",
            "properties": {
                "ip_addresses": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "system.Utilization": {
            "type": "object",
            "properties": {
                "cpu_percent": {
                    "type": "number"
                },
                "disk_details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/system.DiskInfo"
                    }
                },
                "disk_total": {
                    "type": "integer"
                },
                "disk_used": {
                    "type": "integer"
                },
                "load_average1": {
                    "type": "number"
                },
                "load_average15": {
                    "type": "number"
                },
                "load_average5": {
                    "type": "number"
                },
                "memory_total": {
                    "type": "integer"
                },
                "memory_used": {
                    "type": "integer"
                },
                "swap_total": {
                    "type": "integer"
                },
                "swap_used": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "NodeToken": {
            "description": "Supply the node's bearer token from `config.yml` using the `Authorization: Bearer \u003ctoken\u003e` header.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ServerJWT": {
            "description": "Signed JWTs issued by the Panel for server-scoped operations (uploads, downloads, websockets). Pass in the `token` query parameter.",
            "type": "apiKey",
            "name": "token",
            "in": "query"
        }
    }
}